Prelimary magic bitboard work

+++ src/chess_engine/move_gen/basic_sliding.go
@@ -8,18 +8,18 @@ import (
 func slidingRays(ind uint, bb board.Bitboard)  board.Bitboard{
 
 	// get row and column of index
-	row := ind / 8
-	col := ind % 8
+	row := int(ind / 8)
+	col := int(ind % 8)
 	
 	// directions of move
 	directions := [4]int{1, -1, 8, -8}
-	max_dir := [4]uint{7 - col, col, 7 - row, row}
+	max_dir := [4]int{7 - col, col, 7 - row, row}
 	
 	rays := board.Bitboard(0)
 
 	for i, dir := range directions {
 
-		for j := 1; uint(j) <= max_dir[i]; j++ {
+		for j := 1; j <= max_dir[i]; j++ {
 
 			sq := int(ind) + dir * (j)
 			rays |= 1 << sq
@@ -39,19 +39,19 @@ func slidingRays(ind uint, bb board.Bitboard)  board.Bitboard{
 func diagonalRays(ind uint, bb board.Bitboard)  board.Bitboard{
 	
 	// get row and column of index
-	row := ind / 8
-	col := ind % 8
+	row := int(ind / 8)
+	col := int(ind % 8)
 	
 	// directions of move
 	directions := [4]int{9, -9, 7, -7}
-	max_dir := [4]uint{min(7 - col, 7 - row), min(col, row), 
+	max_dir := [4]int{min(7 - col, 7 - row), min(col, row), 
 				min(7 - row, col), min(row, 7 - col)}
 
 	rays := board.Bitboard(0)
 
 	for i, dir := range directions {
 
-		for j := 1; uint(j) <= max_dir[i]; j++ {
+		for j := 1; j <= max_dir[i]; j++ {
 
 			sq := (int(ind)) + dir * (j)
 			rays |= 1 << sq
@@ -67,14 +67,14 @@ func diagonalRays(ind uint, bb board.Bitboard)  board.Bitboard{
 }
 
 
-func min(a, b uint) uint {
+func min(a, b int) int {
     if a < b {
         return a
     }
     return b
 }
 
-func max(a, b uint) uint {
+func max(a, b int) int {
     if a > b {
         return a
     }
